requirements.txt
Flask==3.1.1
tensorflow>=2.15.0
numpy==2.1.3
Pillow==11.3.0
flask-cors==6.0.1